refactor: extract feature-specific modules
