+++ app-expo/app/(tabs)/search/index.tsx
@@ -1,15 +1,14 @@
 import React, { useState, useEffect } from "react";
 import {
-	View,
-	Text,
-	StyleSheet,
-	ScrollView,
-	TouchableOpacity,
-	TextInput,
-	SafeAreaView,
-	ActivityIndicator,
-	FlatList,
-	PanResponder,
+        View,
+        Text,
+        StyleSheet,
+        ScrollView,
+        TouchableOpacity,
+        TextInput,
+        SafeAreaView,
+        ActivityIndicator,
+        FlatList,
 } from "react-native";
 import { Divider } from "react-native-paper";
 import {
@@ -29,79 +28,16 @@ import { SearchParams, SearchLocation, GooglePlacesPrediction } from "@/types/se
 import { useLocationSearch } from "@/hooks/useLocationSearch";
 import { useSnackbar } from "@/contexts/SnackbarProvider";
 import { Card } from "@/components/Card";
-
-const timeSlots = [
-	{ id: "morning", label: "朝食", icon: "🌅" },
-	{ id: "lunch", label: "ランチ", icon: "🌞" },
-	{ id: "dinner", label: "ディナー", icon: "🌙" },
-	{ id: "late_night", label: "夜食", icon: "🌃" },
-] as const;
-
-const sceneOptions = [
-	{ id: "solo", label: "おひとり様", icon: "👤" },
-	{ id: "date", label: "デート", icon: "💕" },
-	{ id: "group", label: "複数人と", icon: "👥" },
-	{ id: "large_group", label: "大人数", icon: "👥👥" },
-	{ id: "tourism", label: "観光", icon: "🌍" },
-] as const;
-
-const moodOptions = [
-	{ id: "hearty", label: "がっつり", icon: "🍖" },
-	{ id: "light", label: "軽めに", icon: "🥗" },
-	{ id: "sweet", label: "甘いもの", icon: "🍰" },
-	{ id: "spicy", label: "辛いもの", icon: "🌶️" },
-	{ id: "healthy", label: "ヘルシー", icon: "🥬" },
-	{ id: "junk", label: "ジャンク", icon: "🍔" },
-	{ id: "alcohol", label: "お酒メイン", icon: "🍺" },
-] as const;
-
-// Distance options in meters
-const distanceOptions = [
-	{ value: 100, label: "100m" },
-	{ value: 300, label: "300m" },
-	{ value: 500, label: "500m" },
-	{ value: 800, label: "800m" },
-	{ value: 1000, label: "1km" },
-	{ value: 2000, label: "2km" },
-	{ value: 3000, label: "3km" },
-	{ value: 5000, label: "5km" },
-	{ value: 10000, label: "10km" },
-	{ value: 15000, label: "15km" },
-	{ value: 20000, label: "20km" },
-];
-
-// Budget options in yen
-const budgetOptions = [
-	{ value: null, label: "下限なし" },
-	{ value: 1000, label: "1,000円" },
-	{ value: 2000, label: "2,000円" },
-	{ value: 3000, label: "3,000円" },
-	{ value: 4000, label: "4,000円" },
-	{ value: 5000, label: "5,000円" },
-	{ value: 6000, label: "6,000円" },
-	{ value: 7000, label: "7,000円" },
-	{ value: 8000, label: "8,000円" },
-	{ value: 9000, label: "9,000円" },
-	{ value: 10000, label: "10,000円" },
-	{ value: 15000, label: "15,000円" },
-	{ value: 20000, label: "20,000円" },
-	{ value: 30000, label: "30,000円" },
-	{ value: 40000, label: "40,000円" },
-	{ value: 50000, label: "50,000円" },
-	{ value: 60000, label: "60,000円" },
-	{ value: 80000, label: "80,000円" },
-	{ value: 100000, label: "100,000円" },
-	{ value: null, label: "上限なし" },
-];
-
-const restrictionOptions = [
-	{ id: "vegetarian", label: "ベジタリアン", icon: "🌱" },
-	{ id: "gluten_free", label: "グルテンフリー", icon: "🌾" },
-	{ id: "dairy_free", label: "乳製品不使用", icon: "🥛" },
-	{ id: "nut_allergy", label: "ナッツアレルギー", icon: "🥜" },
-	{ id: "seafood_allergy", label: "魚介アレルギー", icon: "🐟" },
-	{ id: "halal", label: "ハラール", icon: "🕌" },
-];
+import {
+        timeSlots,
+        sceneOptions,
+        moodOptions,
+        distanceOptions,
+        budgetOptions,
+        restrictionOptions,
+} from "@/features/search/constants";
+import { DistanceSlider } from "@/features/search/components/DistanceSlider";
+import { BudgetSlider } from "@/features/search/components/BudgetSlider";
 
 export default function SearchScreen() {
 	const [location, setLocation] = useState<SearchLocation | null>(null);
@@ -208,104 +144,6 @@ export default function SearchScreen() {
 		} finally {
 			setIsSearching(false);
 		}
-	};
-
-	// Distance slider component
-	const DistanceSlider = () => {
-		const currentIndex = distanceOptions.findIndex((option) => option.value === distance);
-		const sliderWidth = 280;
-		const thumbWidth = 24;
-		const trackWidth = sliderWidth - thumbWidth;
-		const thumbPosition = (currentIndex / (distanceOptions.length - 1)) * trackWidth;
-
-		const panResponder = PanResponder.create({
-			onStartShouldSetPanResponder: () => true,
-			onMoveShouldSetPanResponder: () => true,
-			onPanResponderMove: (evt, gestureState) => {
-				const newPosition = Math.max(0, Math.min(trackWidth, gestureState.moveX - 50));
-				const newIndex = Math.round((newPosition / trackWidth) * (distanceOptions.length - 1));
-				if (newIndex !== currentIndex && newIndex >= 0 && newIndex < distanceOptions.length) {
-					setDistance(distanceOptions[newIndex].value);
-				}
-			},
-		});
-
-		return (
-			<View style={styles.sliderContainer}>
-				<View style={styles.sliderTrack}>
-					<View style={[styles.sliderThumb, { left: thumbPosition }]} {...panResponder.panHandlers} />
-				</View>
-				<View style={styles.sliderLabels}>
-					<Text style={styles.sliderLabelLeft}>近い</Text>
-					<Text style={styles.sliderLabelRight}>遠い</Text>
-				</View>
-			</View>
-		);
-	};
-
-	// Budget range slider component
-	const BudgetSlider = () => {
-		const minIndex = budgetMin === null ? 0 : budgetOptions.findIndex((option) => option.value === budgetMin);
-		const maxIndex =
-			budgetMax === null ? budgetOptions.length - 1 : budgetOptions.findIndex((option) => option.value === budgetMax);
-
-		const sliderWidth = 280;
-		const thumbWidth = 24;
-		const trackWidth = sliderWidth - thumbWidth;
-
-		const minThumbPosition = (minIndex / (budgetOptions.length - 1)) * trackWidth;
-		const maxThumbPosition = (maxIndex / (budgetOptions.length - 1)) * trackWidth;
-
-		const createPanResponder = (isMin: boolean) =>
-			PanResponder.create({
-				onStartShouldSetPanResponder: () => true,
-				onMoveShouldSetPanResponder: () => true,
-				onPanResponderMove: (evt, gestureState) => {
-					const newPosition = Math.max(0, Math.min(trackWidth, gestureState.moveX - 50));
-					const newIndex = Math.round((newPosition / trackWidth) * (budgetOptions.length - 1));
-
-					if (isMin) {
-						if (newIndex <= maxIndex && newIndex >= 0) {
-							setBudgetMin(budgetOptions[newIndex].value);
-						}
-					} else {
-						if (newIndex >= minIndex && newIndex < budgetOptions.length) {
-							setBudgetMax(budgetOptions[newIndex].value);
-						}
-					}
-				},
-			});
-
-		const minPanResponder = createPanResponder(true);
-		const maxPanResponder = createPanResponder(false);
-
-		return (
-			<View style={styles.sliderContainer}>
-				<View style={styles.sliderTrack}>
-					<View
-						style={[
-							styles.rangeTrack,
-							{
-								left: minThumbPosition,
-								width: maxThumbPosition - minThumbPosition + thumbWidth,
-							},
-						]}
-					/>
-					<View
-						style={[styles.sliderThumb, styles.rangeThumbMin, { left: minThumbPosition }]}
-						{...minPanResponder.panHandlers}
-					/>
-					<View
-						style={[styles.sliderThumb, styles.rangeThumbMax, { left: maxThumbPosition }]}
-						{...maxPanResponder.panHandlers}
-					/>
-				</View>
-				<View style={styles.sliderLabels}>
-					<Text style={styles.sliderLabelLeft}>安い</Text>
-					<Text style={styles.sliderLabelRight}>高い</Text>
-				</View>
-			</View>
-		);
 	};
 
 	const formatBudgetRange = () => {
@@ -453,7 +291,7 @@ export default function SearchScreen() {
 								<Text style={styles.sliderValue}>
 									{distanceOptions.find((option) => option.value === distance)?.label}
 								</Text>
-								<DistanceSlider />
+                                                                <DistanceSlider distance={distance} setDistance={setDistance} />
 							</View>
 						</Card>
 
@@ -465,7 +303,12 @@ export default function SearchScreen() {
 							</View>
 							<View style={styles.sliderSection}>
 								<Text style={styles.sliderValue}>{formatBudgetRange()}</Text>
-								<BudgetSlider />
+                                                                <BudgetSlider
+                                                                        budgetMin={budgetMin}
+                                                                        budgetMax={budgetMax}
+                                                                        setBudgetMin={setBudgetMin}
+                                                                        setBudgetMax={setBudgetMax}
+                                                                />
 							</View>
 						</Card>
